Updating links to point to button pages
As a user I would have expected the links to point straight to the mentioned button pages. Currently they lead to the homepages and the user has to search for the button which is a bit inconvenient.

Additionally "Primer Components" redirects to "Primer React" so the link name can probably be changed, right?

Hope I didn't miss a reason as to why the links pointed to the homepages.